fix: attempt to fix drep issue instance
