+++ .changeset/gorgeous-kings-ring.md
@@ -1,0 +1,6 @@
+---
+"@lucid-evolution/core-types": patch
+"@lucid-evolution/lucid": patch
+---
+
+attempt to fix drep issue
