+++ packages/lucid/tsup.config.ts
@@ -6,4 +6,5 @@ export default defineConfig({
   format: ["esm", "cjs"],
   dts: true,
   clean: true,
+  sourcemap: true,
 });
